third_party/blink/web_tests/fast/css-grid-layout/display-grid-set-get-expected.txt

Test that setting and getting display: grid and inline-grid works as expected

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Test getting |display| set through CSS
PASS getComputedStyle(gridElement, '').getPropertyValue('display') is 'grid'
PASS getComputedStyle(inlineGridElement, '').getPropertyValue('display') is 'inline-grid'
PASS getComputedStyle(inlineFloatGridElement, '').getPropertyValue('display') is 'grid'
PASS getComputedStyle(inlineAbsoluteGridElement, '').getPropertyValue('display') is 'grid'


Test getting and setting display through JS
PASS getComputedStyle(element, '').getPropertyValue('display') is 'block'
PASS getComputedStyle(element, '').getPropertyValue('display') is 'grid'
PASS getComputedStyle(element, '').getPropertyValue('display') is 'block'
PASS getComputedStyle(element, '').getPropertyValue('display') is 'inline-grid'
PASS successfullyParsed is true

TEST COMPLETE
